% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{assemble_recipe}
\alias{assemble_recipe}
\title{Assemble a recipe}
\usage{
assemble_recipe(feature_registry, features, df_for_recipe, recipe_roles)
}
\arguments{
\item{feature_registry}{A vector of all the \verb{step_feature_*}s available.}

\item{features}{A character vector of the features needed for your model.}

\item{df_for_recipe}{The data.frame used to build the recipe.}

\item{recipe_roles}{The roles used to build the recipe.}
}
\value{
An unprepped recipe.
}
\description{
Assemble a recipe
}
