% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{build_recipe_call}
\alias{build_recipe_call}
\title{Build a recipe call}
\usage{
build_recipe_call(feature_registry, recipe, feature)
}
\arguments{
\item{feature_registry}{A vector of all the \verb{step_feature_*}s available.}

\item{recipe}{A recipe object.}

\item{feature}{The name of the feature needed.}
}
\value{
An R call.
}
\description{
Build a recipe call
}
