% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_registry_features_needed}
\alias{get_registry_features_needed}
\title{Get registry features needed}
\usage{
get_registry_features_needed(feature_registry, features)
}
\arguments{
\item{feature_registry}{A vector of all the \verb{step_feature_*}s available.}

\item{features}{A character vector of the features needed for your model.}
}
\value{
A character vector.
}
\description{
Get registry features needed
}
