% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_unneeded_recipe_vars}
\alias{get_unneeded_recipe_vars}
\title{Get unneeded recipe vars}
\usage{
get_unneeded_recipe_vars(
  unprepped_recipe_full,
  feature_and_deps_names_used,
  features
)
}
\arguments{
\item{unprepped_recipe_full}{Unprepped recipe, still with all steps included.}

\item{feature_and_deps_names_used}{Character vector of features and dependencies needed.}

\item{features}{A character vector of the features needed for your model.}
}
\value{
A character vector.
}
\description{
Get unneeded recipe vars
}
