% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{init_recipe}
\alias{init_recipe}
\title{Initialize a recipe}
\usage{
init_recipe(df_for_recipe, recipe_roles)
}
\arguments{
\item{df_for_recipe}{The data.frame used to build the recipe.}

\item{recipe_roles}{The roles used to build the recipe.}
}
\value{
An unprepped recipe.
}
\description{
Initialize a recipe
}
