% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{step_feature}
\alias{step_feature}
\title{Create step}
\usage{
step_feature(step_fn, step_name, deps, args)
}
\arguments{
\item{step_fn}{The step function to be used, e.g. recipes::step_mutate}

\item{step_name}{The name of the step/feature.}

\item{deps}{A vector of strings with names of dependencies e.g. c("email").}

\item{args}{Arguments passed to \code{step_fn}.}
}
\value{
A list.
}
\description{
Create step
}
\examples{
step_feature(
    recipes::step_mutate,
    "hour",
    c("created_at"),
    lubridate::hour(created_at))
}
