% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{step_mutate_feat}
\alias{step_mutate_feat}
\title{Create mutate step}
\usage{
step_mutate_feat(step_name, deps, args)
}
\arguments{
\item{step_name}{The name of the step/feature.}

\item{deps}{A vector of strings with names of dependencies e.g. c("email").}

\item{args}{Arguments passed to \code{step_fn}.}
}
\value{
A list.
}
\description{
Create mutate step
}
