% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_registry_features_needed}
\alias{get_registry_features_needed}
\title{Title}
\usage{
get_registry_features_needed(feature_registry, features)
}
\arguments{
\item{features}{}
}
\value{

}
\description{
Title
}
